requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
alpaca-py==0.22.0
annotated-types==0.7.0
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.6.2
charset-normalizer==3.3.2
frozenlist==1.4.1
idna==3.7
msgpack==1.0.8
multidict==6.0.5
numpy==1.26.4
pandas==2.2.2
pydantic==2.7.3
pydantic_core==2.18.4
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.3
six==1.16.0
sseclient-py==1.8.0
typing_extensions==4.12.1
tzdata==2024.1
urllib3==2.2.1
websockets==12.0
yarl==1.9.4
